为了模拟真实情况，将account_的主键名改为bigint类型，使其可以接受雪花主键生成器，并为此做出一系列修改；

+++ src/test/java/indi/mybatis/flying/pojo/Account_.java
@@ -20,8 +20,8 @@ public class Account_ extends PojoSupport<Account_> implements Serializable {
 	private static final long serialVersionUID = 1L;
 
 	@Id
-	@FieldMapperAnnotation(dbFieldName = "id", jdbcType = JdbcType.INTEGER)
-	private Integer id;
+	@FieldMapperAnnotation(dbFieldName = "id", jdbcType = JdbcType.BIGINT)
+	private Long id;
 
 	@FieldMapperAnnotation(dbFieldName = "name", jdbcType = JdbcType.VARCHAR)
 	private java.lang.String name;
@@ -64,11 +64,11 @@ public class Account_ extends PojoSupport<Account_> implements Serializable {
 	private java.util.Collection<LoginLogSource2> loginLogSource2;
 
 	@Override
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
